can now compile to html5, added qualtiy setting

+++ src/Camera/Camera.h
@@ -14,7 +14,7 @@ class Camera
 		void renderMap(Map&);
 	private:
 		void renderRay(Map&, double, double, int, double);
-		void generateBlockLine(std::vector<Line>&, double, double, RayCollision collision, double, double, Map&, double, int);
+		void generateBlockLine(std::vector<Line>&, int, int, RayCollision collision, double, double, Map&, double, int);
 		void renderBlockLines(std::vector<Line>&, int);
 
 };
